backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-admin-rangefilter==0.13.3
django-more-admin-filters==1.13
dotenv==0.9.9
gunicorn==23.0.0
more-itertools==10.8.0
packaging==25.0
psycopg2==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
